Changes to 'whilen...' and 'return' macros
Made changes to the "whilen..." and "return" macros. Changes made to the "whilen..." macros are analogous to the ones that were made to "ifn..." macros. The 'return' macro has been changed so that the carry flag is adjusted JUST BEFORE THE RETURN. This is because "add sp, 0x20" will affect the carry flag.